changes from long to int

+++ src/Replicas/Replica1/MessageComparator.java
@@ -10,8 +10,8 @@ public class MessageComparator implements Comparator<JSONObject> {
     public int compare(JSONObject string, JSONObject t1) {
 
         // Extract sequence numbers
-        long seqNum1 = Integer.parseInt( (String) string.get("sequenceNumber"));
-        long seqNum2 = Integer.parseInt( (String) t1.get("sequenceNumber"));
+        int seqNum1 = Integer.parseInt( (String) string.get("sequenceNumber"));
+        int seqNum2 = Integer.parseInt( (String) t1.get("sequenceNumber"));
 
 
         // Compare sequence numbers
@@ -19,6 +19,6 @@ public class MessageComparator implements Comparator<JSONObject> {
         // if o1 == o2, value will be == 0;
         // if o1 > o2, value will be 1;
 
-        return Long.compare(seqNum1, seqNum2);
+        return Integer.compare(seqNum1, seqNum2);
     }
 }
